Инициализация таблиц и загрузка данных на сервер вынесена в отдельную функцию
