+++ src/usecases/instrument/edit_instrument.ts
@@ -1,9 +1,9 @@
 import Instrument from '../../entities/instrument';
-import instrumentsTableManager from '../../managers/instruments_table_manager';
+import InstrumentsTableManager from '../../managers/instruments_table_manager';
 import getInstument from './get_instrument';
 
 const editInstrument = async (instrument: Instrument) => {
-	await instrumentsTableManager.edit(instrument);
+	await InstrumentsTableManager.edit(instrument);
 
 	return await getInstument(instrument.figi);
 };
