+++ src/usecases/instrument/get_instrument.ts
@@ -1,8 +1,8 @@
 import { INSTRUMENT_NOT_FOUND } from '../../constants/errors';
-import instrumentsTableManager from '../../managers/instruments_table_manager';
+import InstrumentsTableManager from '../../managers/instruments_table_manager';
 
-const getInstument = async (figi: string)=> {
-	const instrument = await instrumentsTableManager.get(figi);
+const getInstument = async (figi: string) => {
+	const instrument = await InstrumentsTableManager.get(figi);
 
 	if (instrument === undefined) {
 		throw Error(INSTRUMENT_NOT_FOUND);
