+++ src/usecases/instrument/search_instrument.ts
@@ -1,12 +1,12 @@
 import Instrument from '../../entities/instrument';
 import InvestmentList from '../../entities/investment_list';
-import instrumentsTableManager from '../../managers/instruments_table_manager';
+import InstrumentsTableManager from '../../managers/instruments_table_manager';
 import { convertToInvestmentList } from '../../utils/convector';
 
 const searchInstrument = async (
 	query: string
 ): Promise<InvestmentList<Instrument>> => {
-	const instruments = await instrumentsTableManager.getAllWhereQueryIsExists(
+	const instruments = await InstrumentsTableManager.getAllWhereQueryIsExists(
 		query
 	);
